Cleanup network files, update gitignore
