+++ .gitignore
@@ -1,10 +1,2 @@
-.vscode/*
-./vscode
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
-*.code-workspace
-
-# Local History for Visual Studio Code
-.history/
+.vscode
+*.exe
